Fix linux gfx issue (#2131)
* #include an absolute path didn't work - because paths were taken to always be relative.

* Fix linux build issue with renderer-shared.h

* Move to .cpp file.

* Move some templates to stop definition order issue on linux.

* Move to header because it's templated.

* Fix warning about GeometryInstanceFlags::Enum flags : 8; being too small by making the enum backed by 8 bits.

* Check if vkDestroySampler function is available.

+++ slang-gfx.h
@@ -909,7 +909,7 @@ public:
     {
         // The enum values are kept consistent with D3D12_RAYTRACING_INSTANCE_FLAGS
         // and VkGeometryInstanceFlagBitsKHR.
-        enum Enum : uint32_t
+        enum Enum : uint8_t
         {
             None = 0,
             TriangleFacingCullDisable = 0x00000001,
